Add shared ResourceDrop type
