+++ Assets/Scripts/Enemies/Enemy.cs
@@ -12,13 +12,6 @@ namespace TimelessEchoes.Enemies
     [RequireComponent(typeof(Health))]
     public class Enemy : MonoBehaviour
     {
-        [System.Serializable]
-        public class ResourceDrop
-        {
-            public Resource resource;
-            public Vector2Int dropRange = new Vector2Int(1, 1);
-            [Range(0f, 1f)] public float dropChance = 1f;
-        }
         [SerializeField] private EnemyStats stats;
         [SerializeField] private Animator animator;
         [SerializeField] private SpriteRenderer spriteRenderer;
