+++ Assets/Scripts/Tasks/MiningTask.cs
@@ -11,13 +11,6 @@ namespace TimelessEchoes.Tasks
     /// </summary>
     public class MiningTask : MonoBehaviour, ITask
     {
-        [System.Serializable]
-        public class ResourceDrop
-        {
-            public Resource resource;
-            public Vector2Int dropRange = new Vector2Int(1, 1);
-            [Range(0f, 1f)] public float dropChance = 1f;
-        }
 
         [SerializeField] private float mineTime = 2f;
         [SerializeField] private SlicedFilledImage progressBar;
